Change IFileSystem to perform read/write operations
Trying to make the FakeFileSystem be writeable when backed by byte arrays or memory stream did not go well.  Byte arrays weren't resizeable.  Memory streams might get closed on accident if exposed to the consumer.

Instead, the FakeFS (and IFS) can just provide helpers since so far we're just reading/writing text.  We can add additional helpers when we need to support binary formats, but this way we have better control of the streams.