Messages for no existing drafts and no exisitng posts (#62)
Display appropriate message on Index and Drafts pages when such posts do not exist.
When there are no published posts, display a welcome message and a link to write a blog post on the Index page. "View and edit your saved drafts here" message is displayed on the Drafts page when there are no saved drafts.