Refactored into ViewModels (#93)
Applied ViewModels to limit editable fields to this context.
Correctly deleted old Posts and Drafts from the data store when posts are updated, as appropriate.
Added Required attributes for NewPostViewModel and EditPostViewModel. Now validating ModelState in Edit.cshtml.cs.
Corrected the logic for setting the PubDate of a post -- should be set only once, when the post is published for the first time.