Excerpt generation for Edited Posts (#159)
Fixed issue: Excerpt is not generated after editing a post.
Now when the blog owner publishes a post, edits the post and deletes the excerpt, and published the post again, an excerpt is generated based on the updated body.
Note: BlogTemplate._1.Tests.Pages.Account.RegisterTests.RegisterUser_FirstUser_LocalRedirect was broken before this branch was created. Should fix in a different pull request.
Wrote a test to confirm that when the excerpt for an edited post is deleted, a new excerpt is generated based on the edited body.